requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0

# 통합 이메일 서비스를 위한 패키지들
pydantic>=2.0.0
python-dateutil>=2.8.0
requests>=2.31.0

# Gmail 연동을 위한 패키지들
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Microsoft Graph API 연동을 위한 패키지들
# requests는 이미 위에 포함됨

# 기타 유틸리티
typing-extensions 